feat: improve the code - remove unnecessary selects file - move chatMessage fixtures in the chat - move itemFavorite fixtures in the favorite - fix typo
